feat(exec): add --last flag to only output the final agent message
This commit introduces a new `--last` flag to the `exec` command.
When this flag is used, only the final message from the agent is printed to stdout.

This is useful for scripting and for situations where the user is only interested in the final output of the agent.

This commit also introduces a new `--output-last-message` flag that allows the user to specify a file where the last message from the agent will be written.

In addition, this commit also includes:

- A version bump for all the crates in the workspace.
- A rename of the CLI from `coco-cli` to `cocos`.
- An increase of the timeout in `core/src/exec.rs` from 20 to 40 minutes.